+++ config/environments/development.rb
@@ -13,6 +13,7 @@ Rails.application.configure do
   config.eager_load = false
 
   # Show full error reports.
+  # Set to 'false' to test dynamic error pages
   config.consider_all_requests_local = true
 
   # Enable/disable caching. By default caching is disabled.
@@ -31,7 +32,7 @@ Rails.application.configure do
 
   # Don't care if the mailer can't send.
   config.action_mailer.raise_delivery_errors = false
-
+  config.action_mailer.default_url_options = { host: 'localhost', port: 3000 }
   config.action_mailer.perform_caching = false
 
   # Print deprecation notices to the Rails logger.