+++ spec/factories/accounts.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+FactoryGirl.define do
+  factory :account do
+    user { build(:user, account: Account.new) }
+    email { Faker::Internet.unique.email(user.name) }
+    password { Faker::Internet.password(8, 128) }
+  end
+end