+++ spec/factories/handles.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+FactoryGirl.define do
+  factory :handle do
+    after(:build) do |handle|
+      handle.profile = build(:user, handle: handle) unless handle.profile
+      handle.identifier =
+        Faker::Internet.user_name(handle.profile.name.first(26).strip, %w[_])
+    end
+  end
+end