+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## v0.2 (August 7, 2017)
+
+**New Features**:
+- Registrations
+- Account Management
+- Sessions (login & logout)
+- User profiles with handles (usernames)
+
+**Minor Changes**:
+- Remove JavaScript from application
+- Add dynamic error pages ([Issue](https://github.com/UpshiftOne/upshift/issues/12))
+- Responsive font-size for headings
+
 ## v0.1.1 (July 30, 2017)
 
 - Fix: Downgrade to Puma 3.8.1 ([Issue](https://github.com/seuros/capistrano-puma/issues/237))